fast_api/collect/amonth_crawrer/requirements.txt
pymongo==4.8.0
kiwipiepy==0.18.0
scikit-learn==1.5.2
sentence-transformers==3.1.1
numpy==1.26.4
joblib==1.4.2
apscheduler==3.10.4
python-dotenv==1.0.1

# 설치 코드 (안될 시 한 번만 실행하세요)
# pip install -r requirements.txt